+++ logger.py
@@ -5,7 +5,6 @@ import colorlog
 import os
 from os import path
 from properties import *
-import app_config
 
 
 if not path.exists('logs'):
@@ -38,5 +37,5 @@ if __name__ != '__main__':
     logs = [f'./logs/{l}' for l in os.listdir('./logs') if l.startswith('session')]
     logs.sort(key=lambda x: os.path.getctime(x))
 
-    if len(logs) > app_config.Config().max_logs:
+    if len(logs) > MAX_SESSION_LOGS:
         os.remove(logs[0])